pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-newsletter-manager"
version = "0.1.0"
description = "Intelligent Gmail newsletter management utility built on gogcli"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Olusola Akinsulere", email = "olusolaakinsulere@gmail.com"}
]
keywords = [
    "gmail",
    "newsletter",
    "email",
    "organization",
    "automation",
    "productivity"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Utilities",
]
dependencies = [
    "rich>=10.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/YOUR-USERNAME/gmail-newsletter-manager"
Documentation = "https://github.com/YOUR-USERNAME/gmail-newsletter-manager/blob/main/README.md"
Repository = "https://github.com/YOUR-USERNAME/gmail-newsletter-manager"
Issues = "https://github.com/YOUR-USERNAME/gmail-newsletter-manager/issues"
Changelog = "https://github.com/YOUR-USERNAME/gmail-newsletter-manager/releases"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.0.270",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
]

[project.scripts]
newsletter-manager = "newsletter_manager.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
target-version = ["py38"]

